Minor Changes
Mostly formatting, but I also:
Updated X.set$ to return the stream/atom like X.over$ does.
Removed a tap wrapping a freeze in UserModel.

This is a safety commit before I try to updated lensedAtom to return
itself when set (rather than returning the value it was set with).